+++ NEWS.md
@@ -24,6 +24,10 @@
 
 - `knitr::knit2html()` still has the `encoding` argument, but it only supports UTF-8 internally.
 
+## MINOR CHANGES
+
+- Changed `tools::texi2dvi()` to `tinytex::latexmk()` for the `tikz` engine to compile TikZ graphics to other foramts (e.g., `svg` or `png`). This requires the **tinytex** package >= v0.10.10: https://github.com/yihui/tinytex.
+
 # CHANGES IN knitr VERSION 1.21
 
 ## NEW FEATURES